% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adamsgld_functions.R
\name{adam_sgld_update}
\alias{adam_sgld_update}
\title{ADAM SGLD update step}
\usage{
adam_sgld_update(
  phi,
  grad_t,
  lr,
  momentum_t,
  V_t,
  beta_1 = 0.9,
  beta_2 = 0.9,
  bias_factor = 4
)
}
\description{
ADAM SGLD update step
}
