% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adamsgld_functions.R
\name{adamsgld_mcmc}
\alias{adamsgld_mcmc}
\title{Main ADAM SGLD function}
\usage{
adamsgld_mcmc(
  y,
  X,
  NNarray,
  locs,
  beta_0,
  covparams0,
  covfun_name = "matern_isotropic",
  lr = 0.001,
  lr_min = 2e-06,
  n_epochs = 100,
  n_batch = 250,
  n_burn = 2000,
  thin = 5,
  beta1_factor = 0.9,
  beta2_factor = 0.99,
  bias_factor = 2,
  lambda = 1e-05,
  covparams_prior_params,
  silent = F,
  initial_V = NULL
)
}
\description{
Main ADAM SGLD function
}
