% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_parms.R
\name{get_start_parms}
\alias{get_start_parms}
\title{get default starting values of covariance parameters}
\usage{
get_start_parms(y, X, locs, covfun_name)
}
\arguments{
\item{y}{response}

\item{X}{design matrix, with the first column being a vector of 1's for the intercept}

\item{locs}{locations}

\item{covfun_name}{string name of covariance function. Options are 'exponential_isotropic', 'matern_isotropic', 'matern15_isotropic'}
}
\description{
get default starting values of covariance parameters
}
