% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{L_t_mult}
\alias{L_t_mult}
\title{Multiply transpose of approximate Cholesky by a vector}
\usage{
L_t_mult(Linv, z, NNarray)
}
\arguments{
\item{Linv}{Entries of the sparse inverse Cholesky factor,
usually the output from \code{\link{vecchia_Linv}}.}

\item{z}{the vector to be multiplied}

\item{NNarray}{Array of nearest neighbors}
}
\value{
the product of the transpose of the Cholesky factor with a vector
}
\description{
Vecchia's approximation implies a sparse approximation to the
inverse Cholesky factor of the covariance matrix. This function
returns the result of multiplying the transpose of the
inverse of that matrix by a vector
(i.e. an approximation to the transpose of the Cholesky factor).
}
