% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Linv_t_mult}
\alias{Linv_t_mult}
\title{Multiply transpose of approximate inverse Cholesky by a vector}
\usage{
Linv_t_mult(Linv, z, NNarray)
}
\arguments{
\item{Linv}{Entries of the sparse inverse Cholesky factor,
usually the output from \code{\link{vecchia_Linv}}.}

\item{z}{the vector to be multiplied}

\item{NNarray}{Array of nearest neighbors}
}
\value{
the product of the transpose of the 
sparse inverse Cholesky factor with a vector
}
\description{
Vecchia's approximation implies a sparse approximation to the
inverse Cholesky factor of the covariance matrix. This function
returns the result of multiplying the transpose of that matrix by a vector.
}
