% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ma_vecchia_profbeta_loglik_grad_info}
\alias{ma_vecchia_profbeta_loglik_grad_info}
\title{Function to evaluate and return the gradient of the loglikelihood}
\usage{
ma_vecchia_profbeta_loglik_grad_info(
  batch_id,
  covparms,
  covfun_name,
  y,
  X,
  current_beta,
  locs,
  NNarray
)
}
\arguments{
\item{batch_id}{Minibatch id}

\item{covparms}{Covariance parameter values}

\item{covfun_name}{Covariance function; 
tested with 'matern_isotropic' and 'exponential_isotropic'}

\item{y}{Response vector}

\item{X}{Design matrix of covariates. Row \code{i} of \code{X} contains
the covariates for the observation at row \code{i} of \code{locs}}

\item{current_beta}{Mean function parameter values}

\item{locs}{Matrix of locations}

\item{NNarray}{Array of nearest neighbors}
}
\value{
A list containing
\itemize{
    \item \code{loglik}: the loglikelihood
    \item \code{grad}: gradient with respect to covariance parameters
    \item \code{info}: Fisher information for covariance parameters
    \item \code{betahat}: profile likelihood estimate of regression coefs
    \item \code{betainfo}: information matrix for \code{betahat}.
    \item \code{grad_beta}: gradient of the loglikelihood with respect to mean parameters
}
The covariance matrix for \code{$betahat} is the inverse of \code{$betainfo}.
}
\description{
Function to evaluate and return the gradient of the loglikelihood
}
