% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgld_functions.R
\name{msgld_mcmc}
\alias{msgld_mcmc}
\title{Main MSGLD function}
\usage{
msgld_mcmc(
  y,
  X,
  NNarray,
  locs,
  beta_0,
  covparams0,
  covfun_name = "matern_isotropic",
  lr = 0.001,
  lr_min = 2e-06,
  n_epochs = 100,
  n_batch = 250,
  n_burn = 2000,
  thin = 5,
  beta_factor = 0.9,
  bias_factor = 1,
  lambda = 1e-05,
  covparams_prior_params,
  silent = F
)
}
\description{
Main MSGLD function
}
