% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgld_functions.R
\name{msgld_update}
\alias{msgld_update}
\title{MSGLD upate step}
\usage{
msgld_update(
  phi,
  grad_t,
  momentum_t,
  beta = 0.9,
  lr,
  bias_factor = 1,
  temperature = 1
)
}
\description{
MSGLD upate step
}
