% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parms_invlink}
\alias{parms_invlink}
\title{Inverse Bijector for the cov parameters
all the parameters are sampled on the log scale}
\usage{
parms_invlink(logparms)
}
\arguments{
\item{logparms}{Parameters on the log scale}
}
\description{
Inverse Bijector for the cov parameters
all the parameters are sampled on the log scale
}
