% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parms_link_grad}
\alias{parms_link_grad}
\title{Bijector gradient for the cov parameters}
\usage{
parms_link_grad(covparms)
}
\arguments{
\item{covparms}{Covariance parameters}
}
\description{
Bijector gradient for the cov parameters
}
