% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgmcmc_functions.R
\name{r_sgrld_step_beta}
\alias{r_sgrld_step_beta}
\title{SGRLD step for betas
Probably not being used anywhere}
\usage{
r_sgrld_step_beta(lr, info, beta_c, beta_hat)
}
\description{
SGRLD step for betas
Probably not being used anywhere
}
