% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reparameterized_quantities}
\alias{reparameterized_quantities}
\title{Function to transform the gradients and fisher information
of the likelihood from the constrained to unconstrained space}
\usage{
reparameterized_quantities(cov_params, grad, info, grad_phi, info_phi)
}
\arguments{
\item{cov_params}{Covariance parameter values}

\item{grad}{gradient with respect to cov_params}

\item{info}{Fisher information for cov_params}

\item{grad_phi}{reference to vector for storing unconstrained gradients}

\item{info_phi}{reference to matrix for unconstrained Fisher info}
}
\description{
Function to transform the gradients and fisher information
of the likelihood from the constrained to unconstrained space
}
