% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_pieces}
\alias{sample_pieces}
\title{Internal stuff with the minibatch}
\usage{
sample_pieces(batch_id, NNarray, covparms, X, y, locs, covfun_name)
}
\arguments{
\item{batch_id}{Minibatch id}

\item{NNarray}{Array of nearest neighbors}

\item{covparms}{Covariance parameter values}

\item{X}{Design matrix of covariates}

\item{y}{Response vector}

\item{locs}{Matrix of locations}

\item{covfun_name}{Covariance function name}
}
\description{
Internal stuff with the minibatch
}
