% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SGRLD_step}
\alias{SGRLD_step}
\title{Function to take an SGDRLD step
for the covariance parameters
This function returns \eqn{( \log(\sigma^2), \log(\alpha), \log(\nu), \log(\tau^2) )_{t+1}}}
\usage{
SGRLD_step(epsilon, info, cov_params, grad)
}
\arguments{
\item{epsilon}{the step size}

\item{info}{the fisher information matrix i.e. preconditionner for the cov parameters}

\item{cov_params}{the current state of the cov parameters}

\item{grad}{sample gradient with respect to \eqn{( \log(\sigma^2), \log(\alpha), \log(\nu), \log(\tau^2) )_{t+1}}}
}
\description{
Function to take an SGDRLD step
for the covariance parameters
This function returns \eqn{( \log(\sigma^2), \log(\alpha), \log(\nu), \log(\tau^2) )_{t+1}}
}
