% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrld_functions.R
\name{update_preconditioner}
\alias{update_preconditioner}
\title{Update preconditioner}
\usage{
update_preconditioner(V, grad, alpha = 0.99, lambda = 1e-05, diag_V = T)
}
\description{
Update preconditioner
}
