% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vecchia_Linv}
\alias{vecchia_Linv}
\title{Entries of inverse Cholesky approximation}
\usage{
vecchia_Linv(covparms, covfun_name, locs, NNarray, start_ind = 1L)
}
\arguments{
\item{covparms}{Covariance parameters}

\item{covfun_name}{Covariance function name}

\item{locs}{Matrix of locations}

\item{NNarray}{Array of nearest neighbors}

\item{start_ind}{Compute entries of Linv only for rows \code{start_ind}
until the last row.}
}
\value{
matrix containing entries of inverse Cholesky
}
\description{
This function returns the entries of the inverse Cholesky
factor of the covariance matrix implied by Vecchia's approximation.
For return matrix \code{Linv}, \code{Linv[i,]} contains
the non-zero entries of row \code{i} of
the inverse Cholesky matrix. The columns of the non-zero entries
are specified in \code{NNarray[i,]}.
}
