% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_grad_logprior}
\alias{beta_grad_logprior}
\title{gradient of log_prior on \eqn{\beta}}
\usage{
beta_grad_logprior(beta, var = 1)
}
\arguments{
\item{beta, }{current value}
}
\value{
vector with same size as \eqn{\beta}
}
\description{
gradient of log_prior on \eqn{\beta}
}
