% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{fit_model_sgmcmc}
\alias{fit_model_sgmcmc}
\title{Spatial SGMCMC using SGRLD}
\usage{
fit_model_sgmcmc(
  y,
  X,
  NNarray,
  locs,
  beta_0,
  covparams0,
  covfun_name = "matern_isotropic",
  algorithm = "SGRLD",
  lr = 0.001,
  lr_min = 2e-06,
  n_epochs = 100,
  n_batch = 250,
  n_burn = 2000,
  thin = 5,
  covparams_prior_params,
  silent = F
)
}
\arguments{
\item{y}{Vector of responses}

\item{X}{Matrix of covariates - usually the first column is going to be 1's for the intercept}

\item{NNarray}{Nearest neighbor object output from GpGp's find_ordered_nn function}

\item{locs}{Matrix of locations with each row corresponding to a location}

\item{beta_0}{Initial values for GP mean parameters}

\item{covfun_name}{Supports "matern_isotropic" and "exponential_isotropic"}

\item{lr}{Learning rate; 1e-3 by default}

\item{lr_min}{Lower bound for learning rate; 2e-6 by default}

\item{n_batch}{Size of each batch}

\item{n_burn}{Burn-in period for MCMC}

\item{thin}{Thin posterior samples}

\item{covparams_prior_params}{Check papaer for the distributions; order same as GpGp}

\item{covparams_0}{Initial values for covariance parameters (same structure as GpGp)}

\item{epochs}{Number of epochs}
}
\value{
A list with 4 components - draws of beta, draws of covariance paramters, trace of loglik, and time
}
\description{
Main function from the paper to draw MCMC samples using SGRLD. Initial values usually obtained
from GpGp. Order of covariance parameters same as GpGp. Thoroughly tested with isotropic_matern.
Might struggle with other covariance structures for the time being.
}
