% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matern_parms_prior_grad}
\alias{matern_parms_prior_grad}
\title{gradient of log_prior for matern cov_parameters}
\usage{
matern_parms_prior_grad(covparms, prior_params)
}
\arguments{
\item{covparms}{\eqn{\sigma^2, \alpha, \nu, \tau^2}
the nugget is \eqn{\sigma^2  \tau^2}
the priors are log-normal (1,1) for the smoothness \eqn{\nu}
Gamma(9,2) for the range \eqn{\alpha}
Gamma(.1,.1) for the spatial variance \eqn{\sigma^2}
and  Gamma(.1, .1) for the scaled nugget}
}
\value{
a vector same length as covparms.
}
\description{
gradient of log_prior for matern cov_parameters
}
