% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parms_invlink_grad}
\alias{parms_invlink_grad}
\title{Inverse Bijector gradient for the cov parameters}
\usage{
parms_invlink_grad(logparms)
}
\description{
Inverse Bijector gradient for the cov parameters
}
