% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parms_link}
\alias{parms_link}
\title{Bijector for the cov parameters
all the parameters are sampled on the log scale}
\usage{
parms_link(covparms)
}
\description{
Bijector for the cov parameters
all the parameters are sampled on the log scale
}
