% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SGRLD_loop}
\alias{SGRLD_loop}
\title{Function to run spSGMCMC for a number of iterations}
\usage{
SGRLD_loop(
  y,
  X,
  NNarray,
  covfun_name,
  locs,
  beta_0,
  covparams0,
  prior_params,
  indexes,
  n_epochs,
  n_batch,
  n_burn,
  lr,
  thin
)
}
\description{
Function to run spSGMCMC for a number of iterations
}
