% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SGRLD_step}
\alias{SGRLD_step}
\title{Function to take an SGDRLD step
for the covariance parameters
This function returns \eqn{( \log(\sigma^2), \log(\alpha), \log(\nu), \log(\tau^2) )_{t+1}}}
\usage{
SGRLD_step(epsilon, info, cov_params, grad)
}
\arguments{
\item{\eqn{\epsilon}}{the step size}

\item{\code{info}}{the fisher information matrix i.e. preconditionner for the cov parameters}
}
\description{
Function to take an SGDRLD step
for the covariance parameters
This function returns \eqn{( \log(\sigma^2), \log(\alpha), \log(\nu), \log(\tau^2) )_{t+1}}
}
