% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transformed_matern_parms_logprior}
\alias{transformed_matern_parms_logprior}
\title{log prior of transformed matern parameters}
\usage{
transformed_matern_parms_logprior(logparms, prior_params)
}
\description{
log prior of transformed matern parameters
}
