% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transformed_matern_parms_logprior_grad}
\alias{transformed_matern_parms_logprior_grad}
\title{log prior gradient of transformed matern parameters}
\usage{
transformed_matern_parms_logprior_grad(logparms, prior_params)
}
\description{
log prior gradient of transformed matern parameters
}
